wip (debugging 1.3B layer5 intermediate activation)

+++ experimental/transformer/include/ops/W8A8B8O8Linear.h
@@ -2,20 +2,21 @@
 
 struct W8A8B8O8Linear_params {
     Matrix3D<int8_t> weight;
-    Matrix3D<int32_t> bias;
+    Matrix3D<int8_t> bias;
     float alpha;
     float beta;
 };
 
-class W8A8B8O8Linear{
-public:
+class W8A8B8O8Linear {
+   public:
     W8A8B8O8Linear(W8A8B8O8Linear_params &params_);
     W8A8B8O8Linear(){};
     void forward(const Matrix3D<int8_t> &x, Matrix3D<int8_t> &output);
     struct matmul_params params;
     float alpha;
     float beta;
-private:
+
+   private:
     std::string profile_name = "W8A8B8O8Linear";
 };
